Update scoring metric explanation and selection in Initial_models.ipynb
The explanation for the chosen scoring metrics in the Initial_models notebook has been expanded. The model's evaluation is now explained to be based on ROC curves and the AUC score. Reasons for choosing the AUC score as a metric, mainly in terms of model performance and handling imbalanced classes, have been added. The score_metric_used variable has also been refined.